Custom Responses

+++ Delivery.Management.Application/Features/DeliveryAllocations/Handlers/Commands/DeleteDeliveryAllocationCommandHandler.cs
@@ -5,7 +5,9 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
 using Delivery.Management.Application.Contracts.Presistence;
+using Delivery.Management.Application.Exceptions;
 using Delivery.Management.Application.Features.DeliveryAllocations.Requests.Commands;
+using Delivery.Management.Domain;
 using MediatR;
 
 namespace Delivery.Management.Application.Features.DeliveryAllocations.Handlers.Commands
@@ -24,6 +26,11 @@ namespace Delivery.Management.Application.Features.DeliveryAllocations.Handlers.
         public async Task<Unit> Handle(DeleteDeliveryAllocationCommand request, CancellationToken cancellationToken)
         {
             var deliveryAllocation = await _deliveryAllocationRepository.GetAsync(request.Id);
+
+            if (deliveryAllocation == null)
+            {
+                throw new NotFoundException(nameof(DeliveryAllocation), request.Id);
+            }
 
             await _deliveryAllocationRepository.DeleteAsync(deliveryAllocation);
