+++ Delivery.Management.Application/Features/DeliveryTypes/Handlers/Commands/DeleteDeliveryTypeCommandHandler.cs
@@ -5,7 +5,9 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
 using Delivery.Management.Application.Contracts.Presistence;
+using Delivery.Management.Application.Exceptions;
 using Delivery.Management.Application.Features.DeliveryTypes.Requests.Commands;
+using Delivery.Management.Domain;
 using MediatR;
 
 namespace Delivery.Management.Application.Features.DeliveryTypes.Handlers.Commands
@@ -24,6 +26,11 @@ namespace Delivery.Management.Application.Features.DeliveryTypes.Handlers.Comman
         public async Task<Unit> Handle(DeleteDeliveryTypeCommand request, CancellationToken cancellationToken)
         {
             var deliveryType = await _deliveryTypeRepository.GetAsync(request.Id);
+
+            if (deliveryType == null)
+            {
+                throw new NotFoundException(nameof(DeliveryType), request.Id);
+            }
 
             await _deliveryTypeRepository.DeleteAsync(deliveryType);
